Split out services into own files/tests
