+++ streaming/service_youtube.py
@@ -3,27 +3,11 @@ import os
 import re
 from difflib import SequenceMatcher
 from enum import IntEnum
-from shutil import copyfile
 
 from googleapiclient.discovery import build
 from googleapiclient.discovery_cache.base import Cache
-from gmusicapi import Mobileclient
-from ytmusicapi import YTMusic as YTMusicClient
 
-from streaming import (
-    StreamingService,
-    StreamingServiceActionNotSupportedError,
-    StreamingServiceTrack,
-)
-
-"""File for Google-owned music clients. Specifically named to avoid conflicts
-with "google" packages.
-"""
-
-
-###############################################################################
-# YouTube
-###############################################################################
+from streaming import StreamingService, StreamingServiceTrack
 
 # Log all request/response headers and bodies
 # import httplib2
@@ -46,6 +30,8 @@ class MemoryCache(Cache):
 
 
 class YouTubeTrack(StreamingServiceTrack):
+    """YouTube song track"""
+
     # Regexp to exclude from searchable video names
     SEARCHABLE_EXCLUDE_EXPRESSIONS = [
         r"\s\(?(HD\s?)?((with |w\/ )?lyrics)?\)?$",  # ()'s
@@ -116,6 +102,8 @@ class YouTubeVideoCategory(IntEnum):
 
 
 class YouTube(StreamingService):
+    """YouTube client"""
+
     VALID_TRACK_URL_PATTERNS = [
         r"https://www.youtube.com/watch\?v=(?P<trackId>[A-Za-z0-9\-\_]+).*",
         r"https://(www.)?youtu.be/(?P<trackId>[A-Za-z0-9\-\_]+).*",
@@ -177,134 +165,6 @@ class YouTube(StreamingService):
         return tracks
 
 
-###############################################################################
-# YouTube Music
-###############################################################################
-
-
-class YTMusicTrack(StreamingServiceTrack):
-    name = None
-    artist = None
-    id = None
-
-    def __init__(self, name, artist, storeId):
-        self.name = name
-        self.artist = artist
-        self.id = storeId
-
-    def share_link(self):
-        """WARNING: This is not going through an API and is subject to break"""
-        return f"https://music.youtube.com/watch?v={self.id}"
-
-
-class YTMusic(StreamingService):
-    VALID_TRACK_URL_PATTERNS = [
-        r"https://music.youtube.com/watch\?v=(?P<trackId>\w+).*",
-    ]
-
-    def __init__(self):
-        self._client = None
-
-    def __enter__(self):
-        self._client = YTMusicClient()
-        return self
-
-    def __exit__(self, *args):
-        pass
-
-    def get_track_from_trackId(self, trackId):
-        # We'll need to switch clients to support this feature
-        raise StreamingServiceActionNotSupportedError()
-
-    def search_tracks(self, q, max_results=None):
-        del max_results  # unused, not available through client
-
-        tracks = []
-        for search_result in self._client.search(q, filter="songs"):
-            track = search_result
-            ytm_track = YTMusicTrack(
-                track["title"], track["artists"][0]["name"], track["videoId"]
-            )
-            tracks.append(ytm_track)
-        return tracks
-
-
-###############################################################################
-# Google Music
-###############################################################################
-
-
-class GMusicTrack(StreamingServiceTrack):
-    name = None
-    artist = None
-    id = None
-
-    def __init__(self, name, artist, storeId):
-        self.name = name
-        self.artist = artist
-        self.id = storeId
-
-    def share_link(self):
-        """WARNING: This is not going through an API and is subject to break"""
-        return f"https://play.google.com/music/m/{self.id}"
-
-
-class GMusic(StreamingService):
-    CRED_FILE = "gmusicapi.cred"
-    VALID_TRACK_URL_PATTERNS = [
-        r"https://play.google.com/music/m/(?P<trackId>\w+)\??.*",
-    ]
-
-    def __init__(self):
-        self._client = Mobileclient()
-
-    def __enter__(self):
-        # AWS / Docker
-        if (
-            os.getenv("LAMBDA_TASK_ROOT")
-            and os.environ["SERVERLESS_STAGE"] != "local"
-        ):
-            cred_path = os.path.join(
-                os.environ["LAMBDA_TASK_ROOT"], "secrets", GMusic.CRED_FILE
-            )
-            # Only /tmp is writable on AWS lambda. gmusicapi needs to write to
-            # the cred file to refresh its tokens.
-            tmp_path = os.path.join("/tmp", GMusic.CRED_FILE)
-            copyfile(cred_path, tmp_path)
-            self._client.oauth_login(
-                Mobileclient.FROM_MAC_ADDRESS, oauth_credentials=tmp_path
-            )
-        # Local
-        else:
-            cred_path = os.path.join(os.getcwd(), "secrets", GMusic.CRED_FILE)
-            self._client.oauth_login(
-                Mobileclient.FROM_MAC_ADDRESS, oauth_credentials=cred_path
-            )
-        return self
-
-    def __exit__(self, *args):
-        self._client.logout()
-
-    def get_track_from_trackId(self, trackId):
-        query_result = self._client.get_track_info(trackId)
-        track = GMusicTrack(
-            query_result["title"],
-            query_result["artist"],
-            query_result["storeId"],
-        )
-        return track
-
-    def search_tracks(self, q, max_results=5):
-        tracks = []
-        for search_result in self._client.search(q, max_results)["song_hits"]:
-            track = search_result["track"]
-            gm_track = GMusicTrack(
-                track["title"], track["artist"], track["storeId"]
-            )
-            tracks.append(gm_track)
-        return tracks
-
-
 if __name__ == "__main__":
     """Integration Tests"""
     with YouTube() as yt:
@@ -313,26 +173,5 @@ if __name__ == "__main__":
         trackId = YouTube.get_trackId_from_url(track.share_link())
         print(trackId)
         track = yt.get_track_from_trackId(trackId)
-        print(track)
-        print(track.share_link())
-    print()
-    with YTMusic() as ytm:
-        track = ytm.search_one_track("G.o.a.t polyphia")
-        print(track)
-        trackId = YTMusic.get_trackId_from_url(track.share_link())
-        print(trackId)
-        try:
-            track = ytm.get_track_from_trackId(trackId)
-            print(track)
-            print(track.share_link())
-        except StreamingServiceActionNotSupportedError:
-            print("StreamingServiceActionNotSupportedError")
-    print()
-    with GMusic() as gm:
-        track = gm.search_one_track("G.o.a.t polyphia")
-        print(track)
-        trackId = GMusic.get_trackId_from_url(track.share_link())
-        print(trackId)
-        track = gm.get_track_from_trackId(trackId)
         print(track)
         print(track.share_link())
